// backend/src/generated/error-codes.ts
// This file was generated by src/cmd/generate-error-map
// based on docs/openapi.yaml & src/cmd/generate-error-map/schema-http-code.json.
// Please DON'T change this file directly!!!
// Apply your changes in docs/openapi.yaml & src/cmd/generate-error-map/schema-http-code.json
// and run `npm run generate-error-map` instead.
export const ErrorCodes = {
    API_VALIDATION_ERROR: 'API_VALIDATION_ERROR',
    UNAUTHORIZED: 'UNAUTHORIZED',
    REQUEST_FORBIDDEN_ERROR: 'REQUEST_FORBIDDEN_ERROR',
    USER_NOT_FOUND: 'USER_NOT_FOUND',
    CONFLICT: 'CONFLICT',
    CODE_EXPIRED: 'CODE_EXPIRED',
    UNPROCESSABLE: 'UNPROCESSABLE',
    INTERNAL_SERVER_ERROR: 'INTERNAL_SERVER_ERROR',
};
export const ErrorCodeMap: { [key: string]: number } = {
    API_VALIDATION_ERROR: 400,
    UNAUTHORIZED: 401,
    REQUEST_FORBIDDEN_ERROR: 403,
    USER_NOT_FOUND: 404,
    CONFLICT: 409,
    CODE_EXPIRED: 410,
    UNPROCESSABLE: 422,
    INTERNAL_SERVER_ERROR: 500,
};